feat: Add rejection section and re-organize landing page
This commit introduces a new section to the landing pages that highlights the problem of impersonal rejection messages in the job application process.

The sections on the responsive landing page have also been re-organized to create a more logical and persuasive user experience.

Additionally, the .gitignore file has been updated to ignore the api/index.js build artifact.